requirements.txt
asgiref==3.11.1
certifi==2026.1.4
charset-normalizer==3.4.4
crispy-bootstrap5==2025.6
Django==6.0.2
django-countries==8.2.0
django-crispy-forms==2.5
django-ranged-response==0.2.0
django-recaptcha==4.1.0
django-simple-captcha==0.6.3
idna==3.11
pillow==12.1.1
psycopg2-binary==2.9.11
pycryptodome==3.23.0
python-dotenv==1.2.1
requests==2.32.5
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
whitenoise==6.11.0
